requirements.txt
alembic==1.15.2
annotated-types==0.7.0
anyio==4.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
fastapi==0.115.12
greenlet==3.1.1
h11==0.14.0
idna==3.10
Mako==1.3.10
MarkupSafe==3.0.2
oandapyV20==0.7.2
psycopg2-binary==2.9.10
pydantic==2.11.3
pydantic_core==2.33.1
python-dotenv==1.1.0
python-multipart==0.0.20
requests==2.32.3
sniffio==1.3.1
SQLAlchemy==2.0.40
starlette==0.46.1
typing-inspection==0.4.0
typing_extensions==4.13.2
urllib3==2.4.0
uvicorn==0.34.0
